pyproject.toml
[project]
name = "living-archive"
version = "0.2.0"
description = "AI-assisted family archive pipeline — photos and documents"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.42",
    "httpx>=0.27",
    "Pillow>=10.0",
    "pydantic>=2.0",
    "pypdf>=4.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
run-slice = "src.run_slice:main"
doc-scan = "src.doc_scan:main"
doc-extract = "src.run_doc_extract:main"
doc-index = "src.doc_index:main"
